Remove old document formatting bits
